Fixed plot2 idx from being float to being int

+++ src/main_application/plot_objects/plot2d/plot2d.h
@@ -17,7 +17,7 @@ public:
         float* p0;
         float* p1;
         float* p2;
-        float* idx_data;
+        int32_t* idx_data;
     };
 
 private:
